refactor: align card view state model
