+++ src/pages/card-view/ui/CardViewPage.tsx
@@ -5,16 +5,16 @@ import { CardView } from "@/features/card-view";
 import { AppLayout } from "@/widgets/app-layout";
 import { RouteNotFound } from "@/widgets/route-not-found";
 
-import { useCardViewContent } from "../model/useCardViewContent";
+import { useCardViewState } from "../model/useCardViewState";
 
 export const CardViewPage: React.FC = () => {
   const params = useParams();
   const cardId = params.id;
   if (cardId == null) throw new Error("invalid card id");
 
-  const content = useCardViewContent(cardId);
+  const state = useCardViewState(cardId);
 
-  if (content == null) {
+  if (state == null) {
     return (
       <RouteNotFound title="Card not found" description="The requested card is unavailable or has been removed." />
     );
@@ -22,7 +22,7 @@ export const CardViewPage: React.FC = () => {
 
   return (
     <AppLayout showHeader>
-      <CardView {...content} />
+      <CardView {...state} />
     </AppLayout>
   );
 };
